Image Service Common, Remove unnecessary contollers, Complete Profile

+++ src/profile/profile.service.ts
@@ -4,16 +4,16 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { UploadApiResponse } from 'cloudinary';
 import mongoose, { Model } from 'mongoose';
-import cloudinary from 'src/cloudinary/cloudinary.config';
 import { Profile } from 'src/schema/Profile.schema';
+import { ImageService } from './../common/image/image.service';
 import { ProfileDto } from './../dto/Profile.dto';
 
 @Injectable()
 export class ProfileService {
   constructor(
     @InjectModel(Profile.name) private profileModel: Model<Profile>,
+    private readonly imageService: ImageService,
   ) {}
 
   async getProfile() {
@@ -25,7 +25,7 @@ export class ProfileService {
     file: Express.Multer.File,
   ): Promise<Profile> {
     if (!file) throw new Error('Profile picture is required');
-    const imageUploadResult = await this.uploadImage(file);
+    const imageUploadResult = await this.imageService.uploadImage(file);
     const data = new this.profileModel({
       ...profileDto,
       profilePicture: imageUploadResult.secure_url,
@@ -50,7 +50,7 @@ export class ProfileService {
     if (!isValid) throw new BadRequestException('Invalid ID');
 
     const singleData = await this.profileModel.findById(id);
-    const profileId = this.getPublicId(singleData.profilePicture);
+    const profileId = this.imageService.getPublicId(singleData.profilePicture);
 
     let profilePicture: string | undefined;
 
@@ -58,10 +58,10 @@ export class ProfileService {
       if (!profileId)
         throw new BadRequestException('Profile picture id is required');
       if (profileId) {
-        await this.deleteImage(profileId);
+        await this.imageService.deleteImage(profileId);
       }
 
-      const uploadResult = await this.uploadImage(file);
+      const uploadResult = await this.imageService.uploadImage(file);
       profilePicture = uploadResult.secure_url;
     }
     const updateProfile = await this.profileModel.findByIdAndUpdate(
@@ -70,7 +70,6 @@ export class ProfileService {
         $set: {
           ...profileDto,
           profilePicture: profilePicture,
-          updatedAt: new Date(),
         },
       },
       { new: true },
@@ -79,47 +78,15 @@ export class ProfileService {
     return updateProfile;
   }
 
-  async uploadImage(file: Express.Multer.File): Promise<UploadApiResponse> {
-    return new Promise((resolve, reject) => {
-      cloudinary.uploader
-        .upload_stream({ folder: 'profile' }, (error, result) => {
-          if (error) reject(error);
-          resolve(result);
-        })
-        .end(file.buffer);
-    });
-  }
-
-  async deleteImage(publicId: string): Promise<void> {
-    try {
-      const result = await cloudinary.uploader.destroy(publicId);
-      return result;
-    } catch (error) {
-      throw new Error(`Failed to delete image: ` + JSON.stringify(error));
-    }
-  }
-
   async deleteProfile(id: string): Promise<void> {
+    const isValid = mongoose.Types.ObjectId.isValid(id);
+    if (!isValid) throw new BadRequestException('Invalid ID');
     const home = await this.profileModel.findById(id);
     if (!home) throw new NotFoundException('Data not found');
-    const profilePictureId = this.getPublicId(home.profilePicture);
+    const profilePictureId = this.imageService.getPublicId(home.profilePicture);
     if (profilePictureId) {
-      await this.deleteImage(profilePictureId);
+      await this.imageService.deleteImage(profilePictureId);
     }
     return this.profileModel.findByIdAndDelete(id);
   }
-
-  getPublicId = (url: string): string => {
-    const parts = url.split('/upload/')[1];
-    if (!parts) return '';
-    const pathSegments = parts.split('/');
-    if (
-      pathSegments[0].startsWith('v') &&
-      !isNaN(Number(pathSegments[0].slice(1)))
-    ) {
-      pathSegments.shift(); // remove the version segment like "v1748281955"
-    }
-    const publicIdWithExt = pathSegments.join('/');
-    return publicIdWithExt.split('.')[0]; // remove extension
-  };
 }
